feat: [FFBV-49] 대시보드 후회지수

+++ src/main/java/com/halggeol/backend/dashboard/dto/DashboardResponseDTO.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 @Builder
 public class DashboardResponseDTO {
 
-    private int avgRegretScore;
+    private Double avgRegretScore;
 //    private List<AssetDTO> assets;
 //    private List<PortfolioDTO> portfolio;
     private List<UnifiedProductRegretRankingResponseDTO> regretRanking;
